+++ src/utils.ts
@@ -1,14 +1,30 @@
-import { debug, warning } from "@actions/core";
 import { info } from "node:console";
 import { createHash } from "node:crypto";
-import { existsSync, readFileSync, renameSync } from "node:fs";
-import { resolve, basename } from "node:path";
+import {
+  copyFileSync,
+  existsSync,
+  readdirSync,
+  readFileSync,
+  renameSync,
+} from "node:fs";
+import { join, resolve, basename } from "node:path";
+import { debug, warning } from "@actions/core";
+import { getExecOutput } from "@actions/exec";
+import { extractZip } from "@actions/tool-cache";
 import { compareVersions, validate } from "compare-versions";
 
 import { getStoredResponse, setStoredResponse } from "./response-storage";
 
 // First Bun version that ships native Windows ARM64 binaries.
 const WINDOWS_ARM64_MIN_VERSION = "1.3.10";
+
+export const exe = (name: string) =>
+  "windows" === getPlatform() ? `${name}.exe` : name;
+
+const normalizeVersion = (v: string) => v.replace(/^v/i, "");
+
+const windows_arm = (os: string, arch: string) =>
+  "windows" === os && ("aarch64" === arch || "arm64" === arch);
 
 export function getCacheKey(url: string): string {
   return `bun-${createHash("sha1").update(url).digest("base64")}`;
@@ -65,7 +81,7 @@ export function addExtension(path: string, ext: string): string {
 
 export function getPlatform(): string {
   const platform = process.platform;
-  if (platform === "win32") return "windows";
+  if ("win32" === platform) return "windows";
 
   return platform;
 }
@@ -83,7 +99,7 @@ export function getArchitecture(
   arch: string,
   version?: string,
 ): string {
-  if (os === "windows" && (arch === "aarch64" || arch === "arm64")) {
+  if (windows_arm(os, arch)) {
     if (!hasNativeWindowsArm64(version)) {
       warning(
         [
@@ -98,7 +114,7 @@ export function getArchitecture(
     }
   }
 
-  if (arch === "arm64") return "aarch64";
+  if ("arm64" === arch) return "aarch64";
   return arch;
 }
 
@@ -109,7 +125,7 @@ export function getAvx2(
   version?: string,
 ): boolean {
   // Workaround for absence of arm64 builds on Windows before 1.3.10 (#130)
-  if (os === "windows" && (arch === "aarch64" || arch === "arm64")) {
+  if (windows_arm(os, arch)) {
     if (!hasNativeWindowsArm64(version)) {
       return false;
     }
@@ -191,6 +207,86 @@ export function readVersionFromFile(
       return output;
     }
   }
+}
+
+export function isVersionMatch(
+  existingRevision: string,
+  requestedVersion?: string,
+): boolean {
+  // If no version specified, default is "latest" - don't match existing
+  if (!requestedVersion) {
+    return false;
+  }
+
+  // Non-pinned versions should never match existing installations
+  if (/^(latest|canary|action)$/i.test(requestedVersion)) {
+    return false;
+  }
+
+  const [existingVersion] = existingRevision.split("+");
+
+  return (
+    normalizeVersion(existingVersion) === normalizeVersion(requestedVersion)
+  );
+}
+
+export async function extractBun(path: string): Promise<string> {
+  for (const entry of readdirSync(path, { withFileTypes: true })) {
+    const { name } = entry;
+    const entryPath = join(path, name);
+    if (entry.isFile()) {
+      if ("bun" === name || "bun.exe" === name) {
+        return entryPath;
+      }
+      if (/^bun.*\.zip/.test(name)) {
+        const extractedPath = await extractZip(entryPath);
+        return extractBun(extractedPath);
+      }
+    }
+    if (/^bun/.test(name) && entry.isDirectory()) {
+      return extractBun(entryPath);
+    }
+  }
+  throw new Error("Could not find executable: bun");
+}
+
+export async function getRevision(exe: string): Promise<string | undefined> {
+  const revision = await getExecOutput(exe, ["--revision"], {
+    ignoreReturnCode: true,
+  });
+  if (0 === revision.exitCode && /^\d+\.\d+\.\d+/.test(revision.stdout)) {
+    return revision.stdout.trim();
+  }
+  const version = await getExecOutput(exe, ["--version"], {
+    ignoreReturnCode: true,
+  });
+  if (0 === version.exitCode && /^\d+\.\d+\.\d+/.test(version.stdout)) {
+    return version.stdout.trim();
+  }
+  return undefined;
+}
+
+/**
+ * Returns the 'Last-Modified' Date only if it is valid and falls
+ * between 'created' and now. Otherwise returns undefined.
+ */
+export function getValidatedLastModified(
+  res: Response,
+  created: Date,
+): Date | undefined {
+  const headerValue = res.headers.get("Last-Modified");
+  if (!headerValue) return undefined;
+
+  const mtime = new Date(headerValue);
+  const mtimeNum = mtime.getTime();
+
+  // 1. Check for 'Invalid Date' (NaN)
+  // 2. Ensure it isn't before the 'created' bound
+  // 3. Ensure it isn't in the future (server clock drift)
+  const isValid =
+    !isNaN(mtimeNum) && mtimeNum > created.getTime() && mtimeNum < Date.now();
+
+  return isValid ? mtime : undefined;
 }
 
 /**
